+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-11-29]
+## [0.0.1 - 2024-12-02]
 ### Added
 - First version of the project
 - Spring Application
@@ -168,3 +168,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Email context project bridgeheads variable
 - Decode Base64 values of email context
 - Enable minus in variables of email context
+- Last document variables in email context